Stop warnings being printed for special recipes
